# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rahul Raj — Portfolio & Systems Architecture API",
    "version": "1.0.0",
    "description": "Production-grade RESTful API documentation for Rahul Raj's modern portfolio platform, built with **Next.js 16**, **TypeScript**, **MongoDB Atlas**, **Cloudinary**, and **AI LLM Orchestration**.\n\n### Key Architectural Pillars\n- **Public Gateway**: Ultra-fast endpoints for portfolio contents, live ATS resume downloads, real-time presence heartbeats, and AI assistant interaction.\n- **Admin Control Studio & Hardware Biometrics**: Role-Based Access Control (RBAC) secured via WebAuthn/FIDO2 hardware biometric passkeys (Fingerprint, TouchID, Windows Hello, FaceID), 4-digit security PIN, master passcode, cryptographic HMAC tokens, or Email OTP resets.\n- **Resilience Engine**: Zero-crash persistence architecture backed by MongoDB Atlas with in-memory caching and real-time health telemetry probes.\n- **Security & Anti-Abuse**: IP-based sliding window rate limiters, anti-bot honeypot traps, and encrypted session cookies.",
    "contact": {
      "name": "Rahul Raj Modi",
      "email": "rahulraj21480@gmail.com",
      "url": "https://github.com/Rahul-2148"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Environment / Origin Server"
    },
    {
      "url": "http://localhost:3000",
...
    }
  ],
...
  "paths": {
...
    "/api/admin/educations": {
      "get": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "Get all education credentials",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of education entries"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "Add or update education credentials",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Education entry saved"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/cloudinary/upload": {
      "post": {
        "tags": [
          "Admin Settings & Media"
        ],
        "summary": "Upload media to Cloudinary CDN",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "description": "Uploads images (DP, project mockups, galleries) or PDF resumes directly to Cloudinary storage.",
        "responses": {
          "200": {
            "description": "Upload successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "url": {
                      "type": "string",
                      "example": "https://res.cloudinary.com/demo/image/upload/v1234/rahul.jpg"
                    },
                    "publicId": {
                      "type": "string",
                      "example": "portfolio/rahul_avatar_123"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/cloudinary/status": {
      "get": {
        "tags": [
          "Admin Settings & Media"
        ],
        "summary": "Get Cloudinary CDN connection status",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cloudinary configuration status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "configured": {
                      "type": "boolean",
                      "example": true
                    },
                    "cloudName": {
                      "type": "string",
                      "example": "rahul-cloud"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminCookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "portfolio_admin_token",
        "description": "Encrypted HTTP-only admin session token generated during login"
      },
      "AdminPasscodeAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-passcode",
        "description": "Secret master passcode configured via ADMIN_PASSCODE environment variable"
      }
    },
...
  }
}